requirements.txt
coreapi==2.3.3
Django==4.2.3
django-filter==23.2
django-phonenumbers==1.0.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
Jinja2==3.1.2
Pillow==10.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
PyYAML==6.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
